pyproject.toml
[tool.poetry]
name = "bcpcrawler"
version = "0.1.0"
description = ""
authors = ["Patrick Owens <1919554+powens@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"
aiohttp = {extras = ["speedups"], version = "^3.8.4"}
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
